requirements.txt
# Base
python-decouple==3.5

# FastApi
fastapi==0.70.0
uvicorn==0.15.0

# Database
sqlalchemy==1.4.26
psycopg2-binary==2.9.1